add menu feature

+++ src/main/java/com/controller/CustomerAddressController.java
@@ -1,6 +1,8 @@
 package com.controller;
 
 import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import org.hibernate.SessionException;
@@ -20,6 +22,7 @@ import com.dto.CustomerAddressDto;
 import com.entity.CustomerAddressEntity;
 import com.entity.CustomerEntity;
 import com.repository.CustomerAddressRepository;
+import com.repository.CustomerRepository;
 
 import jakarta.servlet.http.HttpSession;
 
@@ -29,6 +32,9 @@ public class CustomerAddressController {
 	
 	@Autowired
 	CustomerAddressRepository customerAddressRepository;
+	
+	@Autowired
+	CustomerRepository customerRepository;
 	
 	@PostMapping
 	public ResponseEntity<?> addAddress(@RequestBody CustomerAddressDto customerAddressDto, HttpSession session) {
@@ -76,14 +82,31 @@ public class CustomerAddressController {
 	}
 	
 	@DeleteMapping("/{addressId}")
-	public ResponseEntity<?> deleteAddress(@PathVariable("addressId") Integer addressId) {
-		Optional<CustomerAddressEntity> op = customerAddressRepository.findById(addressId);
-		
-		if(op.isPresent()) {
-			customerAddressRepository.deleteById(addressId);
-			return ResponseEntity.ok("Success");
-		}else {
-			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+	public ResponseEntity<?> deleteAddress(@PathVariable("addressId") Integer addressId, HttpSession session) {
+		try {
+			CustomerEntity customer = (CustomerEntity)session.getAttribute("customer");
+			
+			if(customer == null) {
+				throw new SessionException("Session Exception");
+			}else {
+				Optional<CustomerAddressEntity> op = customerAddressRepository.findById(addressId);
+				
+				if(op.isPresent()) {
+					customerAddressRepository.deleteById(addressId);
+					return ResponseEntity.ok("Success");
+				}else {
+					HashMap<String, String> error = new HashMap<>();
+					error.put("message", "Address Id not Found");
+					return ResponseEntity.status(HttpStatus.NOT_FOUND).body(error);
+				}
+			}
+		}catch(SessionException sessionException) {
+			HashMap<String, String> error = new HashMap<>();
+			error.put("message", "Unauthorized Access! Please Enter Credentials.");
+			error.put("exception", sessionException.getMessage());
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error);
+		}catch(Exception e) {
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
 		}
 	}
 	
@@ -96,6 +119,39 @@ public class CustomerAddressController {
 			return ResponseEntity.ok("Success");
 		}else {
 			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+		}
+	}
+	
+	// read all address
+	@GetMapping("/myaddress/{customerId}")
+	public ResponseEntity<?> getMyAddress(@PathVariable("customerId") Integer customerId, HttpSession session) {
+		try {
+			CustomerEntity customer = (CustomerEntity)session.getAttribute("customer");
+			Optional<CustomerEntity> op = customerRepository.findById(customerId);
+			
+			if(customer == null) {
+				throw new SessionException("Session Exception");
+			}else {
+				if(customerId == customer.getCustomerId()) {			
+					if(op.isPresent()) {
+						List<CustomerAddressEntity> addresses = op.get().getCustomerAddresses();
+						return ResponseEntity.ok(addresses);
+					}else {
+						return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+					}
+				}else {
+					HashMap<String, Object> error = new HashMap<>();
+					error.put("message", "Unauthorized Access! Enter valid Credentials.");
+					return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error);
+				}
+			}
+		}catch(SessionException sessionException) {
+			HashMap<String, String> error = new HashMap<>();
+			error.put("message", "Unauthorized Access! Please Enter Credentials.");
+			error.put("exception", sessionException.getMessage());
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(error);
+		}catch(Exception e) {
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
 		}
 	}
 }
